feat: move sampling to separate module
- Split sampling off from the experiment setup to prepare use of all scenarios
- Use sampling as separate module to create distinct functions for each scenarios
