+++ src/experiments/__init__.py
@@ -1,2 +1,1 @@
 from .experiment_runner import ExperimentRunner
-from .sampling import sample_datasets
